scripts/requirements.txt
# Busca e coleta
requests>=2.31
beautifulsoup4>=4.12
rapidfuzz>=3.0
rispy>=0.8
unidecode>=1.3

# Extração e análise
pdfplumber>=0.11
google-generativeai>=0.8

# Exportação
pandas>=2.0
openpyxl>=3.1

# Configuração
pyyaml>=6.0
